Simplified, started using jquery
bundle.coffee now includes the built-in minified jquery js
Removed a lot of code from strapdown.js, so it no longer sets up script
and theme links, nav bar etc. The aim is to make it more flexible, with
the downside of needing a more fully-formed original html file, the
advantage being that we should be able to easily go to custom bootstrap
3, and also add support for more framework structure in the html - e.g.
a div for TOC extracted from headers.
Converted most of the remaining strapdown.js to use jquery, will
complete this in next commit.